+++ src/index.ts
@@ -37,6 +37,7 @@ export { default as Chip } from '@/package/components/chip.vue';
 export { default as Avatar } from '@/package/components/avatar.vue';
 export { default as DataTable } from '@/package/components/data-table.vue';
 export { default as Slider } from '@/package/components/slider.vue';
+export { default as Stepper } from '@/package/components/stepper.vue';
 export { default as Breadcrumbs } from '@/package/components/breadcrumbs.vue';
 export type { DataTableColumn } from '@/package/components/data-table.vue';
 // config