+++ src/package/themes/default/index.ts
@@ -33,6 +33,7 @@ import badge from './components/badge';
 import chip from './components/chip';
 import avatar from './components/avatar';
 import slider from './components/slider';
+import stepper from './components/stepper';
 
 export default {
     ...base,
@@ -72,5 +73,6 @@ export default {
         chip,
         avatar,
         slider,
+        stepper,
     },
 };